Corrections to new comments
(1) I don't understand what you mean by 'non-optimal auxiliary'. In
your test script, vanderbei_style_test.sage, if you call view(D) after
the line "D.is_optimal())" (which returns True), certainly the aux
problem is optimal -- but it should still be printed with xi as its
objective (which your code does). So why not change 'non-optimal
auxiliary' to just 'auxiliary'?

peijun: (2) The way that InteractiveLPProblemStandardForm.__init__
handles the "objective" argument should be changed.
If the user provides this argument, then we should use it, even if
we're in Vanderbei mode.
So best to change from objective="z" to objective=None, and if it is
None, then compute the defaults, depending on non-Vanderbei mode or
Vanderbei mode.

peijun: (3) The sentence "For educational purposes, style must be
               initialized to vanderbei" does not compute. Just say:
"Style must be one of None (the default) or 'vanderbei’"

peijun: (1 continued): The result of the feasible_dictionary method is
an ordinary primal dictionary, if I'm not mistaken. So why call that an
"optimal auxiliary" dictionary?