change objective variable to dual problem
In InteractiveLPProblem.dual, you could have an optional keyword
argument "objective", defaulting to None.
If the user provides that argument, just use it, regardless of style.
However, if it's not provided (None), then: If style is None, the dual
would just use the objective name of the primal. If style is
'vanderbei', then it would use objective='xi'.

The change to InteractiveLPProblemStandardForm.auxiliary_problem is
similar.