add doc tests to run_dual_simplex_method
add basic_solution to run_dual_simplex_method
add if d.is_optimal()

Revise codes according to comments:

better: use the "dual simplex method”
Dictionaries are not unbounded; problems are.
Besides, it's the dual problem that is unbounded! So the problem is
infeasible.

Need to exit the loop in this case!