Trac #29430: new spkg-configure.m4 for flintqs.
The flintqs package installs the QuadraticSieve executable that is
used (and assumed to be in PATH) by the qsieve() command in SageMath.
This commit adds an spkg-configure.m4 that checks for the existence
of QuadraticSieve in the user's PATH. We don't do any version checks
because there's only one version of flintqs, and its predecessor,

  http://www.friedspace.com/QS/

was last updated in 2007.