Trac #32037: delete a failing doctest.
The one remaining doctest failure after upgrading to lcalc-2.0.1 is
with the output from,

  sage: chi = DirichletGroup(5)[1]
  sage: L = Lfunction_from_character(chi, type="complex")
  sage: L.hardy_z_function(0.4 + 0.3*I)

This was discussed also on Trac #24820, and it seems that neither the
old output nor the new output can be justified. In other words, we
don't know if there's a regression, or if anything needs to be
changed. For now, to proceed with the lcalc upgrade, I have simply
deleted the doctest