Trac #29024: work around Singular issue 1113 in init_libsingular().
When we Singular with dlopen() by name, it is initially unable to
determine the path to its main executable and emits a "BUG" warning.
This is ultimately harmless, but is very scary and appears
unexpectedly whenever Sage is invoked. Here we work around the issue
temporarily by setting SINGULAR_BIN_DIR to the directory containing
the "Singular" executable in the user's PATH.