Trac #29665: add --enable-system-site-packages option to ./configure.
This new option is experimental and disabled by default. When it is
enabled, the string "--system-site-packages" is appended to a new
environment variable SAGE_VENV_FLAGS that is then passed to the
sage-venv command during the build using autoconf substitution.
Thusly the correct setting finds its way into local/pyvenv.cfg.